reorganize  rollout and utils so we don't get import errors
