+++ prob_mbrl/__init__.py
@@ -1,4 +1,3 @@
-from core import custom_pbar, train_regressor
 import utils
 import losses
 import models
