+++ prob_mbrl/utils/__init__.py
@@ -1,6 +1,7 @@
-from core import (iterate_minibatches, plot_sample, plot_mean_var,
-                  plot_trajectories, plot_rollout, batch_jacobian, custom_pbar,
-                  train_regressor)
+from core import (plot_sample, plot_mean_var, plot_trajectories, plot_rollout,
+                  batch_jacobian)
+from train_regressor import train_regressor, iterate_minibatches
+from rollout import rollout
 from experience_dataset import ExperienceDataset
 from apply_controller import apply_controller
 from . import (classproperty, angles)
@@ -8,5 +9,6 @@ from . import (classproperty, angles)
 __all__ = [
     "iterate_minibatches", "plot_sample", "plot_mean_var", "plot_trajectories",
     "plot_rollout", "batch_jacobian", "classproperty", "angles",
-    "ExperienceDataset", "apply_controller", "custom_pbar", "train_regressor"
+    "ExperienceDataset", "apply_controller", "custom_pbar", "train_regressor",
+    "rollout"
 ]
